Fix bug where the controller blocked endpoint deletes
If the owning load balancer somehow was deleted before the endpoint
then the controller would return an error over and over but never
remove its finalizer from the finalizer list so the object couldn't be
deleted.

Now we remove the finalizer much earlier so we don't block deletion
even when things aren't what we expect.