Protect cache updates with a mutex
I was seeing "fatal error: concurrent map writes" at startup time when
a lot of things all happen simultaneously. Two different goroutines
were trying to allocate cache versions and they conflicted when they
tried to update the ServiceGroup's Status. Evidently the caching
controller client returns the same object to different goroutines so
if they try to do something like update a map in that object they can
conflict.

Now we allocate versions in the function that updates the cache (which
is probably cleaner anyway) and we wrap the version allocation and
cache update in a mutex.