Update build scripts
This commit updates the build scripts to look more like our other
projects. While this project was on ice we moved forward in our
understanding of how to build golang projects, especially in terms of
Gitlab CI.

This commit also moves the deployment manifest back from the egw
project to this one so we can have better control over what we
generate, and less coupling between projects.