Add controllers to track Gateway resources
Each GWProxy is a Node in the Envoy cache. The Node ID is the
namespaced name of the resource. GWRoute is the "glue" that binds
GWProxies to services (which map onto Envoy upstream clusters). When a
GWRoute changes it might mean that one or more GWProxy snapshots need
to be updated. GWEndpointSlice holds the IP addresses of the endpoint
pods on the client cluster. For compatibility reasons we translate
GWEndpoint slices into our old-school RemoteEndpoint objects.